server/requirements.txt
tornado==6.4.2
tornado-sqlalchemy==0.7.0
datetime==4.9
python-dateutil==2.9.0.post0
pylint-ignore==2022.1025
marshmallow-sqlalchemy==0.30.0
tornado-prometheus==0.1.2
bcrypt==4.3.0
anytree==2.13.0
alembic==1.15.2
marshmallow<4